requirements.txt
opencv-python == 4.5.3.56
opencv-contrib-python == 4.5.4.58
mediapipe == 0.8.9
av == 8.0.3
pandas == 0.23.4
python-i18n == 0.3.9
playsound == 1.2.2
pyinstaller == 5.0.1
protobuf == 3.20.1